py/requirements.txt
requests
pillow
progress
